+++ cmake/clang-aarch64-toolchain.cmake
@@ -1,0 +1,7 @@
+set(CMAKE_C_COMPILER clang)
+set(CMAKE_CXX_COMPILER clang++)
+set(CMAKE_SYSTEM_PROCESSOR arm64)
+set(CMAKE_C_COMPILER_TARGET aarch64-linux-gnu)
+set(CMAKE_CXX_COMPILER_TARGET aarch64-linux-gnu)
+set(CMAKE_C_FLAGS -isystem/usr/aarch64-linux-gnu/include)
+set(CMAKE_CXX_FLAGS "-isystem/usr/aarch64-linux-gnu/include -isystem/usr/aarch64-linux-gnu/include/c++/10/aarch64-linux-gnu")