+++ src/archive_item.cpp
@@ -35,14 +35,4 @@ void ArchiveItem::adjustTimeOffsets(double diff) {
   m_stop_time_offset += diff;
 }
 
-ArchiveItem& ArchiveItem::operator=(const ArchiveItem& other) {
-  if (this != &other) {
-    this->m_path = other.m_path;
-    this->m_connection_id = other.m_connection_id;
-    this->m_start_time_offset = other.m_start_time_offset;
-    this->m_stop_time_offset = other.m_stop_time_offset;
-  }
-  return *this;
-}
-
 }  // namespace hisui